Moved DefaultParcelableNavTypeSerializer to its own class
